+++ Projects/App/Sources/UI/Place/PlaceList/ViewModel/PlaceListViewModel.swift
@@ -53,22 +53,3 @@ extension PlaceListViewModel: ErrorMapper {
     }
         
 }
-
-protocol ErrorMapper { }
-
-extension ErrorMapper {
-    
-    func errorMessage(for error: NetworkError) -> String {
-        switch error {
-        case .unauthorized, .forbidden:
-            return "권한이 없습니다."
-        case .serverError:
-            return "서버에 문제가 생겼습니다. 다시 시도해주세요."
-        case .unknown:
-            return "알 수 없는 에러..🥲"
-        default:
-            return "문제가 발생했습니다. 제보해주시면 수정하도록 하겠습니다. 😂"
-        }
-    }
-    
-}
