+++ Projects/App/Sources/UI/Review/ViewModel/ReviewRegisterViewModel.swift
@@ -50,7 +50,7 @@ final class ReviewRegisterViewModel {
     
 }
 
-extension ReviewRegisterViewModel {
+extension ReviewRegisterViewModel: ErrorMapper {
     
     func registerReview() {
         useCase.registerReview(placeId: placeId,
@@ -77,19 +77,6 @@ extension ReviewRegisterViewModel {
                                      placeAddress: review.place.address)
             }
             .store(in: &cancelBag)
-    }
-    
-    private func errorMessage(for error: NetworkError) -> String {
-        switch error {
-        case .unauthorized, .forbidden:
-            return "권한이 없습니다."
-        case .serverError:
-            return "서버에 문제가 생겼습니다. 다시 시도해주세요."
-        case .unknown:
-            return "알 수 없는 에러..🥲"
-        default:
-            return "문제가 발생했습니다. 제보해주시면 수정하도록 하겠습니다. 😂"
-        }
     }
     
 }
